+++ project/modules/tap_object_model/image.py
@@ -16,6 +16,7 @@
 
 import uuid
 
+from modules.exceptions import UnexpectedResponseError
 import modules.http_calls.platform.image_factory as image_factory_api
 from ._tap_object_superclass import TapObjectSuperclass
 
@@ -39,4 +40,11 @@ class Image(TapObjectSuperclass):
         return new_image
 
     def delete(self):
-        image_factory_api.delete_image(image_id=self.id)
+        if self._is_deleted is True:
+            return
+
+        try:
+            image_factory_api.delete_image(image_id=self.id)
+            self._set_deleted(True)
+        except UnexpectedResponseError:
+            raise
